rename a file
