+++ crates/app/src/lib.rs
@@ -6,16 +6,16 @@ mod error;
 mod events;
 mod execute;
 mod gas;
+mod outcome;
 mod providers;
 mod query;
 mod shared;
 mod state;
 mod submessage;
 mod traits;
-mod types;
 mod vm;
 
 pub use crate::{
-    app::*, buffer::*, error::*, events::*, execute::*, gas::*, providers::*, query::*, shared::*,
-    state::*, submessage::*, traits::*, types::*, vm::*,
+    app::*, buffer::*, error::*, events::*, execute::*, gas::*, outcome::*, providers::*, query::*,
+    shared::*, state::*, submessage::*, traits::*, vm::*,
 };
