+++ crates/app/src/outcome.rs
